# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the Dora Explorer application.",
        "title": "Dora Explorer API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1/slots": {
            "get": {
                "description": "Returns a list of slots with various filtering options, sorted by slot number descending",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Slot"
                ],
                "summary": "Get filtered slots list",
                "operationId": "getSlots",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by graffiti",
                        "name": "graffiti",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Invert graffiti filter",
                        "name": "graffiti_invert",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by extra data",
                        "name": "extra_data",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Invert extra data filter",
                        "name": "extra_data_invert",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by proposer index",
                        "name": "proposer",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by proposer name",
                        "name": "proposer_name",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Invert proposer filter",
                        "name": "proposer_invert",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Include orphaned blocks (0=exclude, 1=include, 2=only orphaned)",
                        "name": "with_orphaned",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Include missing blocks (0=exclude, 1=include, 2=only missing)",
                        "name": "with_missing",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Minimum sync aggregate participation (0-100)",
                        "name": "min_sync",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Maximum sync aggregate participation (0-100)",
                        "name": "max_sync",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Minimum execution time in ms",
                        "name": "min_exec_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum execution time in ms",
                        "name": "max_exec_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Minimum transaction count",
                        "name": "min_tx_count",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum transaction count",
                        "name": "max_tx_count",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Minimum blob count",
                        "name": "min_blob_count",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum blob count",
                        "name": "max_blob_count",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated list of fork IDs",
                        "name": "fork_ids",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Minimum slot number to return (inclusive)",
                        "name": "min_slot",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum slot number to return (inclusive)",
                        "name": "max_slot",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number for pagination (0-indexed, default 0)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results to return (max 1000, default 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.APISlotsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.APISlotListItem": {
            "type": "object",
            "properties": {
                "attestation_count": {
                    "type": "integer"
                },
                "attester_slashing_count": {
                    "type": "integer"
                },
                "avg_exec_time": {
                    "type": "integer"
                },
                "blob_count": {
                    "type": "integer"
                },
                "block_root": {
                    "type": "string"
                },
                "block_size": {
                    "type": "integer"
                },
                "deposit_count": {
                    "type": "integer"
                },
                "el_extra_data": {
                    "type": "string"
                },
                "epoch": {
                    "type": "integer"
                },
                "eth_block_number": {
                    "type": "integer"
                },
                "eth_transaction_count": {
                    "type": "integer"
                },
                "execution_times": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ExecutionTimeDetail"
                    }
                },
                "exit_count": {
                    "type": "integer"
                },
                "finalized": {
                    "type": "boolean"
                },
                "gas_limit": {
                    "type": "integer"
                },
                "gas_used": {
                    "type": "integer"
                },
                "graffiti": {
                    "type": "string"
                },
                "graffiti_text": {
                    "type": "string"
                },
                "is_mev_block": {
                    "type": "boolean"
                },
                "max_exec_time": {
                    "type": "integer"
                },
                "mev_block_relays": {
                    "type": "string"
                },
                "min_exec_time": {
                    "type": "integer"
                },
                "parent_root": {
                    "type": "string"
                },
                "proposer": {
                    "type": "integer"
                },
                "proposer_name": {
                    "type": "string"
                },
                "proposer_slashing_count": {
                    "type": "integer"
                },
                "recv_delay": {
                    "type": "integer"
                },
                "scheduled": {
                    "type": "boolean"
                },
                "slot": {
                    "type": "integer"
                },
                "state_root": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "sync_aggregate_participation": {
                    "type": "number"
                },
                "time": {
                    "type": "string"
                },
                "with_eth_block": {
                    "type": "boolean"
                }
            }
        },
...
        "api.APISlotsData": {
            "type": "object",
            "properties": {
                "next_page": {
                    "type": "integer"
                },
                "next_slot": {
                    "description": "Use with max_slot for cursor-based pagination",
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "slots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.APISlotListItem"
                    }
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "api.APISlotsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/api.APISlotsData"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.ExecutionTimeDetail": {
            "type": "object",
            "properties": {
                "avg_time": {
                    "type": "integer"
                },
                "client_type": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                },
                "max_time": {
                    "type": "integer"
                },
                "min_time": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}